refactor search/view.py to extract logics into search/logic.py
