+++ webapp/app.py
@@ -12,7 +12,8 @@ from webapp.topics.views import topics
 from webapp.publisher.views import publisher
 from webapp.store.views import store
 from webapp.interfaces.views import interfaces
-from webapp.search.view import search, cache
+from webapp.search.view import search
+from webapp.search.logic import cache
 
 
 app = FlaskBase(
